Refactor into coconut integration tests
