+++ contracts/coconut-bandwidth/src/contract.rs
@@ -72,12 +72,9 @@ pub fn migrate(_deps: DepsMut<'_>, _env: Env, _msg: MigrateMsg) -> Result<Respon
 mod tests {
     use super::*;
     use crate::support::tests::helpers::*;
-    use coconut_bandwidth_contract_common::deposit::DepositData;
     use config::defaults::MIX_DENOM;
+    use cosmwasm_std::coins;
     use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
-    use cosmwasm_std::{coins, Addr};
-    use cw_controllers::AdminError;
-    use cw_multi_test::Executor;
 
     #[test]
     fn initialize_contract() {
@@ -101,91 +98,5 @@ mod tests {
                 .query_balance(env.contract.address, MIX_DENOM.base)
                 .unwrap()]
         );
-    }
-
-    #[test]
-    fn deposit_and_release() {
-        let init_funds = coins(10, MIX_DENOM.base);
-        let deposit_funds = coins(1, MIX_DENOM.base);
-        let release_funds = coins(2, MIX_DENOM.base);
-        let mut app = mock_app(&init_funds);
-        let multisig_addr = String::from(MULTISIG_CONTRACT);
-        let pool_addr = String::from(POOL_CONTRACT);
-        let random_addr = String::from(RANDOM_ADDRESS);
-
-        let code_id = app.store_code(contract_bandwidth());
-        let msg = InstantiateMsg {
-            multisig_addr: multisig_addr.clone(),
-            pool_addr: pool_addr.clone(),
-        };
-        let contract_addr = app
-            .instantiate_contract(
-                code_id,
-                Addr::unchecked(OWNER),
-                &msg,
-                &[],
-                "bandwidth",
-                None,
-            )
-            .unwrap();
-
-        let msg = ExecuteMsg::DepositFunds {
-            data: DepositData::new(
-                String::from("info"),
-                String::from("id"),
-                String::from("enc"),
-            ),
-        };
-        app.execute_contract(
-            Addr::unchecked(OWNER),
-            contract_addr.clone(),
-            &msg,
-            &deposit_funds,
-        )
-        .unwrap();
-
-        // try to release more then it's in the contract
-        let msg = ExecuteMsg::ReleaseFunds {
-            funds: release_funds[0].clone(),
-        };
-        let err = app
-            .execute_contract(
-                Addr::unchecked(multisig_addr.clone()),
-                contract_addr.clone(),
-                &msg,
-                &[],
-            )
-            .unwrap_err();
-        assert_eq!(ContractError::NotEnoughFunds, err.downcast().unwrap());
-
-        // try to call release from non-admin
-        let msg = ExecuteMsg::ReleaseFunds {
-            funds: deposit_funds[0].clone(),
-        };
-        let err = app
-            .execute_contract(
-                Addr::unchecked(random_addr),
-                contract_addr.clone(),
-                &msg,
-                &[],
-            )
-            .unwrap_err();
-        assert_eq!(
-            ContractError::Admin(AdminError::NotAdmin {}),
-            err.downcast().unwrap()
-        );
-
-        let msg = ExecuteMsg::ReleaseFunds {
-            funds: deposit_funds[0].clone(),
-        };
-        app.execute_contract(
-            Addr::unchecked(multisig_addr),
-            contract_addr.clone(),
-            &msg,
-            &[],
-        )
-        .unwrap();
-        let pool_bal = app.wrap().query_balance(pool_addr, MIX_DENOM.base).unwrap();
-        assert_eq!(pool_bal, deposit_funds[0]);
     }
 }
